+++ prep/assets/clubs.py
@@ -94,6 +94,3 @@ class ClubsProcessor(BaseProcessor):
   def process(self):
     self.prep_dfs = [self.process_segment(prep_df) for prep_df in self.raw_dfs]
     self.prep_df = pandas.concat(self.prep_dfs, axis=0).drop_duplicates(subset='club_id', keep='last')
-
-  def get_validations(self):
-    return []
