+++ prep/config.yml
@@ -14,6 +14,8 @@ assets:
   - name: competitions
     class: CompetitionsProcessor
 settings:
+  # seasons: [2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021]
+  seasons: [2019, 2020, 2021]
   competition_codes: [
       'ES1', 'GB1', 'L1', 'IT1', 'FR1', 'GR1', 'PO1', 
       'BE1', 'UKR1', 'BE1', 'RU1', 'DK1', 'SC1', 'TR1',
@@ -22,3 +24,18 @@ settings:
       'RUSS', 'GRP', 'SFA', 'CIT', 'DKP', 'FRCH', 'UKRS',
       'SCI', 'CL', 'EL', 'KLUB', 'ECLQ', 'CLQ', 'ELQ', 'USC'
     ]
+  logging:
+    version: 1
+    loggers:
+      asset:
+        handlers: [console]
+        level: DEBUG
+    handlers:
+      console:
+        class: logging.StreamHandler
+        level: DEBUG
+        stream: ext://sys.stdout
+        formatter: main
+    formatters:
+      main:
+        format: '%(module)s - %(message)s'
